Removing older code and methods
